fix: refactoring pagination to avoid visual overflow
